chore: some tweaks (#20)
Includes additional paths in in the git ignore file: `*.test` and `test-mergesort` to ignore the binaries produced
by the makefile. Changes C standard to gnu11 to allow for non-POSIX GNU extensions `random()` to be used.
Also reorganises the unit test sources and their corresponding recipes in the makefile; test sources have been moved to the `tests` subdirectory.